Map is being sent and rendered on the client rendering the tiles on the client requires the client to have all the necessary tiles in the assets/tiles folder.

+++ shared/src/net/gtamps/shared/serializer/communication/data/DataMap.java
@@ -74,7 +74,7 @@ public final class DataMap extends AbstractSendableData<DataMap> implements Iter
 	}
 
 	public int getInt(final String key) throws NoSuchElementException {
-		return (int)(long) getValue(key, Long.class);
+		return (int) getValue(key, Integer.class);
 	}
 
 	public float getFloat(final String key) throws NoSuchElementException {
